test-requirements.txt
pytest
pytest-cov
pytest-mock
pytest-pep8
vcrpy